requirements.txt
aiofiles==23.1.0
aiogram==3.0.0b7
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.10.4
APScheduler==3.10.1
async-timeout==4.0.2
asyncpg==0.27.0
attrs==23.1.0
Babel==2.9.1
certifi==2022.12.7
charset-normalizer==3.1.0
Deprecated==1.2.13
distro==1.8.0
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
magic-filter==1.0.9
Mako==1.2.4
MarkupSafe==2.1.2
multidict==6.0.4
netaddr==0.8.0
pydantic==1.10.7
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
redis==4.5.5
requests==2.30.0
six==1.16.0
SQLAlchemy==2.0.10
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==2.0.2
var-dump==1.2
wrapt==1.15.0
yarl==1.9.1
yookassa==2.3.5
zope.event==4.6
zope.interface==6.0
